+++ src/Reader/FilterHandler/ILikeHandler/BaseILikeHandler.php
@@ -1,0 +1,30 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Data\Cycle\Reader\FilterHandler\ILikeHandler;
+
+use Cycle\Database\Injection\Fragment;
+use Yiisoft\Data\Cycle\Exception\UnexpectedFilterException;
+use Yiisoft\Data\Reader\Filter\ILike;
+use Yiisoft\Data\Reader\FilterHandlerInterface;
+use Yiisoft\Data\Cycle\Reader\QueryBuilderFilterHandler;
+use Yiisoft\Data\Reader\FilterInterface;
+
+abstract class BaseILikeHandler implements QueryBuilderFilterHandler, FilterHandlerInterface
+{
+    public function getFilterClass(): string
+    {
+        return ILike::class;
+    }
+
+    public function getAsWhereArguments(FilterInterface $filter, array $handlers): array
+    {
+        if (!$filter instanceof ILike) {
+            throw new UnexpectedFilterException(ILike::class, $filter::class);
+        }
+
+        /** @infection-ignore-all mb_strtoupper -> strtoupper */
+        return [new Fragment("UPPER({$filter->getField()})"), 'like', mb_strtoupper($filter->getValue())];
+    }
+}
