+++ src/Reader/FilterHandler/ILikeHandler/PostgresILikeHandler.php
@@ -5,18 +5,11 @@ declare(strict_types=1);
 namespace Yiisoft\Data\Cycle\Reader\FilterHandler\ILikeHandler;
 
 use Yiisoft\Data\Cycle\Exception\UnexpectedFilterException;
-use Yiisoft\Data\Cycle\Reader\QueryBuilderFilterHandler;
 use Yiisoft\Data\Reader\Filter\ILike;
-use Yiisoft\Data\Reader\FilterHandlerInterface;
 use Yiisoft\Data\Reader\FilterInterface;
 
-final class PostgresILikeHandler implements QueryBuilderFilterHandler, FilterHandlerInterface
+final class PostgresILikeHandler extends BaseILikeHandler
 {
-    public function getFilterClass(): string
-    {
-        return ILike::class;
-    }
-
     public function getAsWhereArguments(FilterInterface $filter, array $handlers): array
     {
         if (!$filter instanceof ILike) {
